requirements.txt
python-telegram-bot
python-dotenv
openai
selenium
telethon
undetected-chromedriver>=3.5.5
selenium-stealth
webdriver-manager
setuptools
pytz
fastapi==0.103.2
uvicorn[standard]
aiomysql
asyncpraw
praw
requests
snscrape
sphinx
sphinx_rtd_theme
unittest-xml-reporting
discord.py
matrix-nio
pytest>=7.0.0
pytest-asyncio>=0.21.0
python-multipart
